rename primary component files within a folder to index.js
