+++ src/components/PageSections/Menu/index.js
@@ -3,7 +3,7 @@ import { FormattedMessage, defineMessages, injectIntl } from 'react-intl';
 import PropTypes from 'prop-types'
 
 import I18nNavLink from 'components/I18nLinks/I18nNavLink';
-import ExternalLink from 'components/ExternalLink/ExternalLink';
+import ExternalLink from 'components/ExternalLink';
 import MenuItem from 'components/PageSections/Menu/MenuItem';
 
 import CloseSVG from 'js/svg/Close';
