+++ src/components/StaticMap/index.js
