+++ src/stories/LanguageSelectBar.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import LanguageSelectBar from 'components/PageSections/LanguageSelectBar/LanguageSelectBar';
+import LanguageSelectBar from 'components/PageSections/LanguageSelectBar';
 import { SUPPORTED_LANGUAGES } from 'js/i18n/constants';
 
 storiesOf('LanguageSelectBar', module)
