+++ src/stories/PageBanner.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import PageBanner from 'components/PageBanner/PageBanner';
+import PageBanner from 'components/PageBanner';
 import dummyImage from 'stories/static_data/images/pagebanner_test.jpg';
 
 storiesOf('PageBanner', module)
