+++ src/stories/SectionHeader.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import SectionHeader from 'components/SectionHeader/SectionHeader';
+import SectionHeader from 'components/SectionHeader';
 import ArrowRight from 'js/svg/ArrowRight';
 
 storiesOf('SectionHeader', module)
