+++ src/stories/Steps.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import Steps from 'components/Steps/Steps';
+import Steps from 'components/Steps';
 
 const stepsHtml = `
   <ul>
